app/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.25.6
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.1
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
discord==2.3.2
discord.py==2.3.2
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
filelock==3.13.4
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
ipykernel==6.29.2
ipython==8.22.1
jedi==0.19.1
joblib==1.4.0
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.4
packaging==23.2
pandas==2.2.2
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
propcache==0.2.0
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rank-bm25==0.2.2
regex==2024.4.28
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
stack-data==0.6.3
starlette==0.36.3
threadpoolctl==3.4.0
tokenizers==0.19.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.9.4
